Reformat rust runtime tests
