+++ runtimes/rust/tests/basic.rs
@@ -15,348 +15,31 @@
 use exonum::{
     blockchain::{
         config::{GenesisConfig, GenesisConfigBuilder, InstanceInitParams},
-        Blockchain, BlockchainBuilder, BlockchainMut, Schema as CoreSchema,
+        Blockchain, BlockchainBuilder, BlockchainMut,
     },
-    crypto::Hash,
-    helpers::{generate_testnet_config, Height, ValidatorId},
-    merkledb::{access::AccessExt, BinaryValue, ObjectHash, Patch, Snapshot, SystemSchema},
-    messages::{AnyTx, Verified},
-    runtime::{
-        migrations::{InitMigrationError, MigrationScript},
-        versioning::Version,
-        CallInfo, CoreError, ExecutionContext, InstanceSpec, InstanceStatus, Mailbox, Runtime,
-        WellKnownRuntime,
-    },
+    helpers::{generate_testnet_config, Height},
+    merkledb::{access::AccessExt, BinaryValue, SystemSchema},
+    runtime::{CoreError, InstanceStatus},
 };
 use exonum_derive::{exonum_interface, BinaryValue, ServiceDispatcher, ServiceFactory};
-use futures::{sync::mpsc, Future};
+use futures::sync::mpsc;
 use pretty_assertions::assert_eq;
 use serde_derive::*;
 
-use std::{
-    collections::BTreeMap,
-    sync::{Arc, Mutex},
-};
+use std::collections::BTreeMap;
 
 use exonum_rust_runtime::{
-    ArtifactId, CallContext, Caller, CommonError, DefaultInstance, ErrorMatch, ExecutionError,
-    InstanceId, RustRuntime, RustRuntimeBuilder, Service, ServiceFactory, SnapshotExt,
-    SUPERVISOR_INSTANCE_ID,
+    CallContext, Caller, CommonError, DefaultInstance, ErrorMatch, ExecutionError,
+    RustRuntimeBuilder, Service, ServiceFactory, SnapshotExt,
 };
 
-fn add_transactions_into_pool(
-    blockchain: &mut BlockchainMut,
-    txs: Vec<Verified<AnyTx>>,
-) -> Vec<Hash> {
-    blockchain
-        .merge({
-            let fork = blockchain.fork();
-            let mut schema = CoreSchema::new(&fork);
-            for tx in txs.clone() {
-                schema.add_transaction_into_pool(tx);
-            }
-            fork.into_patch()
-        })
-        .unwrap();
+use self::inspected::{
+    create_block_with_transactions, execute_transaction,
+    DeployArtifact, EventsHandle, Inspected, RuntimeEvent, StartService, StopService,
+    ToySupervisor, ToySupervisorService,
+};
 
-    txs.into_iter().map(|x| x.object_hash()).collect()
-}
-
-fn execute_transaction(
-    blockchain: &mut BlockchainMut,
-    tx: Verified<AnyTx>,
-) -> Result<(), ExecutionError> {
-    let tx_hash = tx.object_hash();
-
-    let (block_hash, patch) = create_block_with_transactions(blockchain, vec![tx]);
-    blockchain
-        .commit(patch, block_hash, vec![], &mut BTreeMap::new())
-        .unwrap();
-
-    let snapshot = blockchain.snapshot();
-    let schema = CoreSchema::new(&snapshot);
-    let location = schema.transactions_locations().get(&tx_hash).unwrap();
-    schema.transaction_result(location).unwrap()
-}
-
-fn create_block_with_transactions(
-    blockchain: &mut BlockchainMut,
-    transactions: Vec<Verified<AnyTx>>,
-) -> (Hash, Patch) {
-    let tx_hashes = add_transactions_into_pool(blockchain, transactions);
-
-    let height = {
-        let snapshot = blockchain.snapshot();
-        CoreSchema::new(&snapshot).next_height()
-    };
-
-    blockchain.create_patch(
-        ValidatorId::zero().into(),
-        height,
-        &tx_hashes,
-        &mut BTreeMap::new(),
-    )
-}
-
-#[derive(Debug, PartialEq)]
-enum RuntimeEvent {
-    Initialize,
-    Resume,
-    BeforeTransactions(Height, InstanceId),
-    DeployArtifact(ArtifactId, Vec<u8>),
-    StartAdding(InstanceSpec, Vec<u8>),
-    Migrate(ArtifactId, Version),
-    StartResuming(InstanceSpec, Vec<u8>),
-    CommitService(Height, InstanceSpec, InstanceStatus),
-    AfterTransactions(Height, InstanceId),
-    AfterCommit(Height),
-    Shutdown,
-}
-
-#[derive(Debug, Clone, Default)]
-struct EventsHandle(Arc<Mutex<Vec<RuntimeEvent>>>);
-
-impl EventsHandle {
-    fn push(&self, event: RuntimeEvent) {
-        self.0.lock().unwrap().push(event);
-    }
-
-    #[must_use]
-    fn take(&self) -> Vec<RuntimeEvent> {
-        self.0.lock().unwrap().drain(..).collect()
-    }
-}
-
-/// Test runtime wrapper logging all the events (as `RuntimeEvent`) happening within it.
-/// For service hooks the logged height is the height of the block **being processed**.
-/// Other than logging, it just redirects all the calls to the inner runtime.
-/// Used to test that workflow invariants are respected.
-#[derive(Debug)]
-struct Inspected<T> {
-    runtime: T,
-    events: EventsHandle,
-}
-
-impl<T: Runtime> Inspected<T> {
-    fn new(runtime: T) -> Self {
-        Self {
-            runtime,
-            events: Default::default(),
-        }
-    }
-}
-
-impl<T: Runtime> Runtime for Inspected<T> {
-    fn initialize(&mut self, blockchain: &Blockchain) {
-        self.events.push(RuntimeEvent::Initialize);
-        self.runtime.initialize(blockchain)
-    }
-
-    fn on_resume(&mut self) {
-        self.events.push(RuntimeEvent::Resume);
-        self.runtime.on_resume()
-    }
-
-    fn deploy_artifact(
-        &mut self,
-        test_service_artifact: ArtifactId,
-        deploy_spec: Vec<u8>,
-    ) -> Box<dyn Future<Item = (), Error = ExecutionError>> {
-        self.events.push(RuntimeEvent::DeployArtifact(
-            test_service_artifact.clone(),
-            deploy_spec.clone(),
-        ));
-        self.runtime
-            .deploy_artifact(test_service_artifact, deploy_spec)
-    }
-
-    fn is_artifact_deployed(&self, id: &ArtifactId) -> bool {
-        self.runtime.is_artifact_deployed(id)
-    }
-
-    fn initiate_adding_service(
-        &self,
-        context: ExecutionContext<'_>,
-        spec: &InstanceSpec,
-        parameters: Vec<u8>,
-    ) -> Result<(), ExecutionError> {
-        self.events.push(RuntimeEvent::StartAdding(
-            spec.to_owned(),
-            parameters.clone(),
-        ));
-        self.runtime
-            .initiate_adding_service(context, spec, parameters)
-    }
-
-    fn initiate_resuming_service(
-        &self,
-        context: ExecutionContext<'_>,
-        spec: &InstanceSpec,
-        parameters: Vec<u8>,
-    ) -> Result<(), ExecutionError> {
-        self.events.push(RuntimeEvent::StartResuming(
-            spec.to_owned(),
-            parameters.clone(),
-        ));
-        self.runtime
-            .initiate_resuming_service(context, spec, parameters)
-    }
-
-    fn update_service_status(
-        &mut self,
-        snapshot: &dyn Snapshot,
-        spec: &InstanceSpec,
-        status: &InstanceStatus,
-    ) {
-        snapshot
-            .for_dispatcher()
-            .get_instance(spec.id)
-            .expect("Service instance should be exists");
-
-        let core_schema = CoreSchema::new(snapshot);
-        let height = core_schema.next_height();
-
-        self.events.push(RuntimeEvent::CommitService(
-            height,
-            spec.to_owned(),
-            status.to_owned(),
-        ));
-        self.runtime.update_service_status(snapshot, spec, status)
-    }
-
-    fn migrate(
-        &self,
-        new_artifact: &ArtifactId,
-        data_version: &Version,
-    ) -> Result<Option<MigrationScript>, InitMigrationError> {
-        self.events.push(RuntimeEvent::Migrate(
-            new_artifact.to_owned(),
-            data_version.clone(),
-        ));
-        self.runtime.migrate(new_artifact, data_version)
-    }
-
-    fn execute(
-        &self,
-        context: ExecutionContext<'_>,
-        call_info: &CallInfo,
-        arguments: &[u8],
-    ) -> Result<(), ExecutionError> {
-        self.runtime.execute(context, call_info, arguments)
-    }
-
-    fn before_transactions(
-        &self,
-        context: ExecutionContext<'_>,
-        instance_id: u32,
-    ) -> Result<(), ExecutionError> {
-        let height = CoreSchema::new(&*context.fork).next_height();
-        self.events
-            .push(RuntimeEvent::BeforeTransactions(height, instance_id));
-        self.runtime.after_transactions(context, instance_id)
-    }
-
-    fn after_transactions(
-        &self,
-        context: ExecutionContext<'_>,
-        instance_id: u32,
-    ) -> Result<(), ExecutionError> {
-        let schema = CoreSchema::new(&*context.fork);
-        let height = schema.next_height();
-        self.events
-            .push(RuntimeEvent::AfterTransactions(height, instance_id));
-        self.runtime.after_transactions(context, instance_id)
-    }
-
-    fn after_commit(&mut self, snapshot: &dyn Snapshot, mailbox: &mut Mailbox) {
-        let height = CoreSchema::new(snapshot).next_height();
-        self.events.push(RuntimeEvent::AfterCommit(height));
-        self.runtime.after_commit(snapshot, mailbox);
-    }
-
-    fn shutdown(&mut self) {
-        self.events.push(RuntimeEvent::Shutdown);
-        self.runtime.shutdown();
-    }
-}
-
-impl WellKnownRuntime for Inspected<RustRuntime> {
-    const ID: u32 = RustRuntime::ID;
-}
-
-#[derive(Debug, Serialize, Deserialize, BinaryValue)]
-#[binary_value(codec = "bincode")]
-struct DeployArtifact {
-    test_service_artifact: ArtifactId,
-    spec: Vec<u8>,
-}
-
-#[derive(Debug, Serialize, Deserialize, BinaryValue)]
-#[binary_value(codec = "bincode")]
-struct StartService {
-    spec: InstanceSpec,
-    constructor: Vec<u8>,
-}
-
-#[derive(Debug, Serialize, Deserialize, BinaryValue)]
-#[binary_value(codec = "bincode")]
-struct StopService {
-    instance_id: InstanceId,
-}
-
-#[exonum_interface]
-trait ToySupervisor<Ctx> {
-    type Output;
-
-    fn deploy_artifact(&self, context: Ctx, request: DeployArtifact) -> Self::Output;
-    fn start_service(&self, context: Ctx, request: StartService) -> Self::Output;
-    fn stop_service(&self, context: Ctx, request: StopService) -> Self::Output;
-}
-
-#[derive(Debug, ServiceFactory, ServiceDispatcher)]
-#[service_dispatcher(implements("ToySupervisor"))]
-#[service_factory(artifact_name = "toy_supervisor", artifact_version = "0.1.0")]
-struct ToySupervisorService;
-
-impl ToySupervisor<CallContext<'_>> for ToySupervisorService {
-    type Output = Result<(), ExecutionError>;
-
-    fn deploy_artifact(
-        &self,
-        mut context: CallContext<'_>,
-        request: DeployArtifact,
-    ) -> Self::Output {
-        context
-            .supervisor_extensions()
-            .start_artifact_registration(request.test_service_artifact, request.spec);
-        Ok(())
-    }
-
-    fn start_service(&self, mut context: CallContext<'_>, request: StartService) -> Self::Output {
-        context
-            .supervisor_extensions()
-            .initiate_adding_service(request.spec, request.constructor)
-    }
-
-    fn stop_service(&self, mut context: CallContext<'_>, request: StopService) -> Self::Output {
-        context
-            .supervisor_extensions()
-            .initiate_stopping_service(request.instance_id)
-    }
-}
-
-impl Service for ToySupervisorService {}
-
-impl DefaultInstance for ToySupervisorService {
-    const INSTANCE_ID: u32 = SUPERVISOR_INSTANCE_ID;
-    const INSTANCE_NAME: &'static str = "supervisor";
-
-    fn default_instance(&self) -> InstanceInitParams {
-        self.artifact_id()
-            .into_default_instance(Self::INSTANCE_ID, Self::INSTANCE_NAME)
-            .with_constructor(Init::default())
-    }
-}
+mod inspected;
 
 #[derive(Debug, Clone, Serialize, Deserialize, BinaryValue)]
 #[binary_value(codec = "bincode")]
